% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_info_function.R
\name{mp_info}
\alias{mp_info}
\title{mp_info}
\usage{
mp_info(constituency)
}
\arguments{
\item{constituency}{constituency name}
}
\value{
Dataframe of current MP information for constituency
}
\description{
Provides current data on an MP for a particular constituency
}
